Move file handling to application service level

+++ lib/application/ExportImportApplicationService.ts
@@ -1,5 +1,6 @@
-import TimeLog from "../domain/TimeLog";
 import TimeLogRepository from "../infrastructure/TimeLogRepository";
+import TimeLogsFile from "../infrastructure/TimeLogsFile";
+import { format } from "date-fns";
 
 export default class ExportImportApplicationService {
   private readonly timeLogRepository: TimeLogRepository;
@@ -8,13 +9,18 @@ export default class ExportImportApplicationService {
     this.timeLogRepository = new TimeLogRepository();
   }
 
-  async getAllTimeLogs() {
-    return await this.timeLogRepository.all();
+  async importFile(file: File) {
+    const timeLogs = await new TimeLogsFile().read(file);
+    if (timeLogs.length > 0) {
+      await this.timeLogRepository.deleteAll();
+      await this.timeLogRepository.saveAll(timeLogs);
+    }
   }
 
-  async import(timeLogs: TimeLog[]) {
-    await this.timeLogRepository.deleteAll();
-    await this.timeLogRepository.saveAll(timeLogs);
+  async exportAllTimeLogs() {
+    const timeLogs = await this.timeLogRepository.all();
+    const fileName = `${format(new Date(), "yyyy-MM-dd_HH-mm")}_overtimer.json`;
+    await new TimeLogsFile().write(fileName, timeLogs);
   }
 }
